fix(docusign): update endpoint context typing to support client context object
